Moved PipeConfig-related classes to the model/pipe-config subdirectory
