+++ src/grandorgue/model/pipe-config/GOPipeConfig.cpp
