+++ src/grandorgue/model/pipe-config/GOPipeConfig.h
