+++ src/grandorgue/model/pipe-config/GOPipeConfigNode.cpp
