+++ src/grandorgue/model/pipe-config/GOPipeConfigNode.h
