+++ src/grandorgue/model/pipe-config/GOPipeConfigTreeNode.cpp
