+++ src/grandorgue/model/pipe-config/GOPipeConfigTreeNode.h
