+++ src/grandorgue/model/pipe-config/GOPipeUpdateCallback.h
